+++ ladim/state.py
@@ -1,12 +1,54 @@
+import pandas as pd
+import numpy as np
 from .model import Model, Module
 
 
 class State(Module):
     def __init__(self, model: Model):
+        """
+        The state module contains static and dynamic particle properties
+
+        The other modules interact with the state module mostly through
+        the getitem and setitem methods. For instance, to increase the
+        depth of all particles by 1, use
+
+        >>> model.state['Z'] += 1
+
+        :param model: Parent model
+        """
         super().__init__(model)
 
     @property
     def size(self):
+        """
+        Current number of particles
+        """
+        raise NotImplementedError
+
+    @property
+    def released(self):
+        """
+        Total number of released particles
+        """
+        raise NotImplementedError
+
+    def append(self, particles: dict):
+        """
+        Add new particles
+
+        Missing variables are assigned a default value of 0.
+
+        :param particles: A mapping from variable names to values
+        """
+        raise NotImplementedError
+
+    def remove(self, particles):
+        """
+        Remove particles
+
+        :param particles: Boolean index of particles to remove
+        :return:
+        """
         raise NotImplementedError
 
     def __getitem__(self, item):
@@ -19,83 +61,47 @@ class State(Module):
         return self.size
 
 
-class WarmStartState(State):
-    def __init__(self, model: Model, **conf):
+class DynamicState(State):
+    def __init__(self, model: Model):
         super().__init__(model)
-        from .legacy.state import State as LegacyState
-        self._state = LegacyState(model, **conf)
 
-    def update(self):
-        self._state.update()
+        self._num_released = 0
+        self._varnames = set()
 
-    def append(self, *args):
-        self._state.append(*args)
+        self._data = pd.DataFrame()
 
-    def warm_start(self, config, grid):
-        self._state.warm_start(self, config, grid)
+    @property
+    def released(self):
+        return self._num_released
+
+    def append(self, particles: dict):
+        num_new_particles = next(len(v) for v in particles.values())
+        particles['pid'] = np.arange(num_new_particles) + self._num_released
+        particles['alive'] = np.ones(num_new_particles, dtype=bool)
+
+        new_particles = pd.DataFrame(data=particles)
+        self._data = pd.concat(
+            objs=[self._data, new_particles],
+            axis='index',
+            ignore_index=True,
+            join='outer',
+        )
+
+        self._num_released += num_new_particles
+
+    def remove(self, particles):
+        if not np.any(particles):
+            return
+
+        keep = ~particles
+        self._data = self._data.iloc[keep]
 
     @property
     def size(self):
-        return len(self._state)
-
-    @property
-    def timestep(self):
-        return self._state.timestep
-
-    @property
-    def timestamp(self):
-        return self._state.timestamp
-
-    @property
-    def ibm_variables(self):
-        return self._state.ibm_variables
-
-    @property
-    def instance_variables(self):
-        return self._state.instance_variables
+        return len(self._data)
 
     def __getitem__(self, item):
-        return self._state[item]
+        return self._data[item].values
 
     def __setitem__(self, item, value):
-        self._state[item] = value
-
-    @property
-    def pid(self):
-        return self._state.pid
-
-    @pid.setter
-    def pid(self, value):
-        self._state.pid = value
-
-    @property
-    def X(self):
-        return self._state.X
-
-    @X.setter
-    def X(self, value):
-        self._state.X = value
-
-    @property
-    def Y(self):
-        return self._state.Y
-
-    @Y.setter
-    def Y(self, value):
-        self._state.Y = value
-
-    @property
-    def Z(self):
-        return self._state.Z
-
-    @Z.setter
-    def Z(self, value):
-        self._state.Z = value
-
-    @property
-    def alive(self):
-        return self._state.alive
-
-    @alive.setter
-    def alive(self, value):
-        self._state.alive = value
+        self._data[item] = value